Clean up PenetrationAsPointPair tests
1. Move valuable tests from proximity_engine_test into their rightful
   home penetration_as_point_pair_callback_test.cc
   - (ProximityEngineTests, Issue10577Regression_Osculation)
   - (BoxPenetrationTest, *)
2. Delete redundant tests (tests that duplicate tests on the Callback or
   are simply bitrotted):
  - The SimplePenetrationTest test harness.
  - (SimplePenetrationTest, PenetrationDynamicAndAnchored)
  - (SimplePenetrationTest, PenetrationDynamicAndDynamicSingleSource)
  - (SimplePenetrationTest, HasCollisionsDynamicAndAnchored)
  - (SimplePenetrationTest, HasCollisionsDynamicAndDynamicSingleSource)
  - (SimplePenetrationTest, WithCollisionFilters)
  - (ProximityEngineTests, ExpressionUnsupported)
  - (ProximityEngineCollisionTest, SpherePunchThroughBox) (an anacrhonism)
3. Respell (SimplePenetrationTest, HasCollisionsAnchoredInterference)
   - This had nothing to do with SimplePenetrationTest. It has been
     elevated to ProximityEngineTests.
4. Reframing existing tests
  - Clarify (ProximityEngineTests, AnchoredBroadPhaseInitialization) and
    use a cheaper query.
5. Create a single test for ProximityEngine's unique responsibilites for
   calls to ComputePointPairPenetration().
   Created: (ProximityEngineQueryTest, ComputePointPairPenetration) and the
   corresponding test suite to frame all such tests.
   Removed the tests that individually exercised aspects of the overall
   test:
    - (ProximityEngineTests, PenetrationSingleAnchored)
    - (ProximityEngineTests, PenetrationMultipleAnchored)
    - (ProximityEngineTests, PenetrationAsPointPairResultOrdering)
