+++ geometry/proximity/test/fcl_utilities.h
@@ -6,6 +6,7 @@
 
 #include "drake/geometry/geometry_ids.h"
 #include "drake/geometry/shape_specification.h"
+#include "drake/math/rigid_transform.h"
 
 namespace drake {
 namespace geometry {
@@ -13,10 +14,11 @@ namespace internal {
 
 // Creates an fcl::CollisionObjectd for a given Drake Shape, stamping the
 // geometry id into the FCL object's user data (as required by
-// the various Callback types).
-std::unique_ptr<fcl::CollisionObjectd> MakeFclObject(const Shape& shape,
-                                                     GeometryId id,
-                                                     bool is_dynamic);
+// the various Callback types) and setting the fcl object's pose to the given
+// pose.
+std::unique_ptr<fcl::CollisionObjectd> MakeFclObject(
+    const Shape& shape, GeometryId id, bool is_dynamic,
+    const math::RigidTransformd& X_WG = math::RigidTransformd::Identity());
 
 }  // namespace internal
 }  // namespace geometry
